Add files via upload
Chatbot with Rule-Based Response
In this project, we have implemented a rule-based chatbot. The chatbot is designed to respond to user queries and engage in conversations based on predefined rules and patterns. It can be a great starting point for building a simple chatbot with specific responses for common user inputs. This is done by spacy library